The View-related interfaces are now deprecated.
The terminology was updated and left in for reference, but it's obvious that no single set of interfaces covers the model-view technique described in the comments, as it's flexible enough to warrant different representations depending on the particular app, emphasizing aspects of it and de-emphasizing others. This will be turned into articles and these files will then be removed from this repository.